backend/requirements.txt
absl-py==2.1.0
altair==5.3.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.28.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==6.2.6
certifi==2024.7.4
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.1
colorama==0.4.6
comm==0.2.2
comtypes==1.2.0
contourpy==1.1.0
cryptography==46.0.5
cycler==0.11.0
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
Django==5.0
django-crispy-forms==2.1
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.24
fastjsonschema==2.20.0
filelock==3.13.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.42.0
fqdn==1.5.1
frozendict==2.4.7
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
gotrue==2.9.1
grpcio==1.65.1
h11==0.14.0
h2==4.3.0
h5py==3.11.0
hf-xet==1.4.2
hpack==4.1.0
html5lib==1.1
httpcore==1.0.5
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.7.2
hyperframe==6.1.0
idna==3.7
imageio==2.34.2
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.13.0
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keras==3.4.1
kiwisolver==1.4.4
lazy_loader==0.4
libclang==18.1.1
lxml==6.0.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mmh3==5.2.1
multidict==6.7.1
multitasking==0.0.12
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.25.2
openai==1.12.0
opencv-contrib-python==4.10.0.84
opencv-python==4.8.0.74
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
orjson==3.11.7
overrides==7.7.0
packaging==23.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
peewee==4.0.2
phonenumbers==8.13.31
php==1.2.1
Pillow==10.1.0
platformdirs==4.2.0
plotly==5.24.1
postgrest==2.28.3
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.4.1
protobuf==4.25.4
psutil==6.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.3.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.18.0
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pypiwin32==223
pyroaring==1.0.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pyttsx3==2.90
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
realtime==2.28.3
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rich-toolkit==0.19.7
rpds-py==0.19.1
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.37.2
storage3==2.28.3
streamlit==1.36.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
supafunc==0.3.3
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.7.2
tinycss2==1.3.0
tokenizers==0.22.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.24.1
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
ujson==5.12.0
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.29.0
virtualenv==20.25.1
watchdog==4.0.1
watchfiles==1.1.1
Wave==0.0.2
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
yarl==1.23.0
yfinance==0.2.37
zstandard==0.25.0
